feat: make components more independent, more customizable
