--- src/__tests__/index.test.tsx
@@ -1,1 +1,0 @@
-it.todo('write a test');
