+++ src/components/extras/BrushPreview.tsx
