+++ src/components/extras/icons/Delete.tsx
