+++ src/components/extras/icons/Palette.tsx
