--- src/components/index.ts
@@ -1,5 +1,0 @@
-export { default as Button } from './Button';
-export { default as ColorPicker } from './colorPicker/ColorPicker';
-export { default as BrushPreview } from './renderer/BrushPreview';
-export { default as SVGRenderer } from './renderer/SVGRenderer';
-export { default as BrushProperties } from './brushProperties/BrushProperties';
