+++ src/constants.ts
@@ -80,5 +80,13 @@ export const DEFAULT_COLORS = [colors, grayscale];
 export const DEFAULT_THICKNESS = 3;
 export const DEFAULT_OPACITY = 1;
 export const DEFAULT_TOOL = DrawingTool.Brush;
+export const DEFAULT_BRUSH_PREVIEW = 'stroke';
+export const DEFAULT_OPACITY_STEP = 0.1;
+export const DEFAULT_THICKNESS_STEP = 1;
+export const DEFAULT_THICKNESS_MIN = 5;
+export const DEFAULT_THICKNESS_MAX = 35;
+export const DEFAULT_SLIDER_COLOR = '#000';
+export const DEFAULT_DELETE_BUTTON_COLOR = '#81090A';
+export const DEFAULT_OTHER_BUTTONS_COLOR = '#DDD';
 export const DEFAULT_ERASER_SIZE = 5;
 export const SLIDERS_HEIGHT = 80;
