+++ src/index.tsx
@@ -1,9 +1,25 @@
 export {
-  default as Draw,
-  DrawRef,
-  DrawProps,
-  DrawInitialValues,
+  default as Canvas,
+  CanvasRef,
+  CanvasProps,
   SimplifyOptions,
-} from './Draw';
+} from './components/canvas/Canvas';
+
+export {
+  default as BottomSection,
+  BottomSectionProps,
+} from './components/extras/BottomSection';
+
+export {
+  default as BrushProperties,
+  BrushPropertiesProps,
+} from './components/extras/brushProperties/BrushProperties';
+
+export {
+  default as ColorPicker,
+  ColorPickerProps,
+} from './components/extras/brushProperties/colorPicker/ColorPicker';
+
 export * from './types';
-export { createSVGPath } from './utils';
+export * from './components/canvas/utils';
+export * from './components/extras/brushProperties/colorPicker/utils';
