--- src/utils/index.ts
@@ -1,2 +1,0 @@
-export { isBright } from './colors';
-export { createSVGPath } from './svg';
